+++ mobile/android/android-components/components/browser/state/src/test/java/mozilla/components/browser/state/engine/EngineObserverTest.kt
@@ -38,6 +38,7 @@ import mozilla.components.concept.engine.mediasession.MediaSession
 import mozilla.components.concept.engine.permission.PermissionRequest
 import mozilla.components.concept.engine.prompt.PromptRequest
 import mozilla.components.concept.engine.shopping.ProductAnalysis
+import mozilla.components.concept.engine.shopping.ProductRecommendation
 import mozilla.components.concept.engine.window.WindowRequest
 import mozilla.components.concept.fetch.Response
 import mozilla.components.support.test.libstate.ext.waitUntilIdle
@@ -84,6 +85,13 @@ class EngineObserverTest {
                 onResult: (ProductAnalysis) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
+
+            override fun requestProductRecommendations(
+                url: String,
+                onResult: (List<ProductRecommendation>) -> Unit,
+                onException: (Throwable) -> Unit,
+            ) {}
+
             override fun findAll(text: String) {}
             override fun findNext(forward: Boolean) {}
             override fun clearFindMatches() {}
@@ -154,6 +162,13 @@ class EngineObserverTest {
                 onResult: (ProductAnalysis) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}
+
+            override fun requestProductRecommendations(
+                url: String,
+                onResult: (List<ProductRecommendation>) -> Unit,
+                onException: (Throwable) -> Unit,
+            ) {}
+
             override fun findAll(text: String) {}
             override fun findNext(forward: Boolean) {}
             override fun clearFindMatches() {}
@@ -213,6 +228,12 @@ class EngineObserverTest {
             ) {}
             override fun checkForPdfViewer(
                 onResult: (Boolean) -> Unit,
+                onException: (Throwable) -> Unit,
+            ) {}
+
+            override fun requestProductRecommendations(
+                url: String,
+                onResult: (List<ProductRecommendation>) -> Unit,
                 onException: (Throwable) -> Unit,
             ) {}